Allocate empty tensor instead of empty_like in binary ops, fix pow
We should allocate an empty tensor as a result tensor when performing
binary ops. Currently some ops use `empty_like(self)` as the initial
result tensor before passing it into TensorIterator. This is not very
efficient because TensorIterator may resize the tensor due to
broadcasting, causing more memory allocation. By using an empty tensor
as the result tensor, we only need to allocate/resize memory once as
opposed to twice.

Also fixes https://github.com/pytorch/pytorch/issues/26495. The bug
there is that the implementation of `pow` is missing a resize in one
case.

Test Plan:
- new test
- run tests

ghstack-source-id: 0705c58c7e406cdc04a30a4585572f0a93b58606
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26498