Renames `tensor.renamed -> rename`, `tensor.names_ -> rename_`
This makes the naming more consistent with PyTorch's API. The original
concern was that `tensor.rename` might make the operation seem like it
is in-place. However, we have many "verb" APIs: `tensor.add(other)`, for
example, doesn't add other to tensor in-place, but `tensor.add_(other)`
does.

`tensor.rename_` does exactly the same place as `tensor.rename`, but
in-place.

Test Plan:
- [namedtensor ci]

ghstack-source-id: e9aa0ce403d92d23f0ba8f63e34a8c71aef327ac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/26548